Split off RubinBandpass into its own file, bandpass.py
