+++ imsim/__init__.py
@@ -18,6 +18,7 @@ from .templates import *
 from .photon_ops import *
 from .flat import *
 from .sky_model import *
+from .bandpass import *
 from .telescope_loader import *
 from .lsst_image import *
 from .checkpoint import *
