+++ imsim/batoid_wcs.py
@@ -15,7 +15,7 @@ from galsim.config import WCSBuilder, RegisterWCSType, GetInputObj
 from lsst.obs.lsst.translators.lsst import SIMONYI_LOCATION as RUBIN_LOC
 from .camera import get_camera
 from .utils import pixel_to_focal, focal_to_pixel
-from .sky_model import RubinBandpass
+from .bandpass import RubinBandpass
 
 # There are 5 coordinate systems to handle.  In order:
 #   ICRF (rc, dc)
